chore(tooling): local pipeline-hardening (golangci-lint, coverage floor, pre-commit) (#33)
Add local, pinned developer tooling and static-analysis gates. Pure hygiene;
no runtime behaviour change.

- .golangci.yml: curated linter set (errcheck, gosimple, govet, ineffassign,
  staticcheck, unconvert, bodyclose, noctx, goimports, gofmt); Go 1.22; pinned
  to golangci-lint v1.61.0. misspell/unused/errorlint parked pending review.
- Makefile: pinned tool versions (golangci-lint v1.61.0, govulncheck v1.1.4,
  gosec v2.20.0) + targets tools/lint/vuln/sec/cover/gate-full. gate-full =
  gate + lint + cover. vuln and sec are advisory (documented) pending a
  toolchain-patch decision and a gosec triage respectively.
- .pre-commit-config.yaml: gofmt/vet/golangci-lint/check-pii local hooks +
  existing gitleaks.
- Fix three zero-behaviour lint findings: smtp errcheck (explicit discard of
  client.Quit), slack ineffassign (drop dead init), cron unconvert (redundant
  uint64 conversion).

Gate: gofmt/vet/build/test -race green; golangci-lint clean; coverage 75.1%
(floor 74%); make smoke 17/17.

+++ internal/scheduler/cron.go
@@ -72,7 +72,7 @@ func ParseCron(expr string) (*CronSchedule, error) {
 	}
 
 	c := &CronSchedule{
-		minute:   uint64(min),
+		minute:   min,
 		hour:     uint32(hr),
 		dom:      uint32(dom),
 		month:    uint16(mon),
